+++ CHANGELOG.md
@@ -1784,3 +1784,4 @@
 - [2026-06-05 17:31] fix(security): fix memory leak by cancelling security subscription
 - [2026-06-05 18:01] feat(auth): add bi-directional event bus handler for auth
 - [2026-06-05 18:07] style(analytics): harmonize light and dark theme palette for analytics
+- [2026-06-05 18:13] perf(network): apply RepaintBoundary on animated network widgets